# 01/web-programming/mastersplat/mastersplat_app/server/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "MastersPlat",
    "description": "Plataforma de agregação e descoberta de dissertações de mestrado. Ao agregar um conjunto de fontes de informação (Data Providers), é possível pesquisar e consultar todas as teses inseridas na aplicação e associadas a cada uma das fontes de informação. "
  },
  "host": "pw-mastersplat.herokuapp.com",
  "basePath": "/api/v1",
...
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/publications": {
      "get": {
        "tags": [
          "publications"
        ],
        "summary": "Todas as publicações",
        "description": "Retorna uma lista de Publicações.  É possivel, através dos parametros definir o limite e o offset (skip) dos items, podendo assim paginar o resultado, aplicar projeções e ordenações aos dados para obter resultado pretendido. É possivél também aplicar filtros avançados aos dados através de parametros em query. Ex:. “field>0” (https://www.npmjs.com/package/api-query-params).",
        "operationId": "getAllPublications",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "número de items a visualizar",
            "type": "string"
          },
          {
            "name": "skip",
            "in": "query",
            "description": "offset aos items",
            "type": "string"
          },
          {
            "name": "sort",
            "in": "query",
            "description": "ordernar o resultado por um/varios atridutos",
            "type": "string"
          },
          {
            "name": "fields",
            "in": "query",
            "description": "informação que pretende obter (field1,field2,...)",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "total": {
                  "type": "integer",
                  "format": "int64"
                },
                "limit": {
                  "type": "integer",
                  "format": "int64"
                },
                "skip": {
                  "type": "integer",
                  "format": "int64"
                },
                "page": {
                  "type": "integer",
                  "format": "int64"
                },
                "pages": {
                  "type": "integer",
                  "format": "int64"
                },
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Publication"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error on get all Publications!"
          }
        }
      }
    },
    "/publications/{id}": {
      "get": {
        "tags": [
          "publications"
        ],
        "summary": "Informação de uma publicação",
        "description": "Para um id, retorna informação da publicação associada.",
        "operationId": "getPublication",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id da publicação desejado",
            "required": true,
            "type": "string"
          },
          {
            "name": "fields",
            "in": "query",
            "description": "informação que pretende obter (field1,field2,...)",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Get Publication successfully!",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "$ref": "#/definitions/Publication"
                }
              }
            }
          },
          "400": {
            "description": "Invalid Publication ID!"
          },
          "404": {
            "description": "Publication not found!"
          },
          "500": {
            "description": "Error when try to find Publication!"
          }
        }
      }
    },
    "/logs": {
      "get": {
        "tags": [
          "logs"
        ],
        "summary": "Todos os Logs",
        "description": "Retorna uma lista de Logs. É possivel, através dos parametros definir o limite e o offset (skip) dos items, podendo assim paginar o resultado, aplicar projeções e ordenações aos dados para obter resultado pretendido. É possivél também aplicar filtros avançados aos dados através de parametros em query. Ex:. “field>0” (https://www.npmjs.com/package/api-query-params).",
        "operationId": "getAllLogs",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "número de items a visualizar",
            "type": "string"
          },
          {
            "name": "skip",
            "in": "query",
            "description": "offset aos items",
            "type": "string"
          },
          {
            "name": "sort",
            "in": "query",
            "description": "ordernar o resultado por um/varios atridutos",
            "type": "string"
          },
          {
            "name": "fields",
            "in": "query",
            "description": "informação que pretende obter (field1,field2,...)",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "total": {
                  "type": "integer",
                  "format": "int64"
                },
                "limit": {
                  "type": "integer",
                  "format": "int64"
                },
                "skip": {
                  "type": "integer",
                  "format": "int64"
                },
                "page": {
                  "type": "integer",
                  "format": "int64"
                },
                "pages": {
                  "type": "integer",
                  "format": "int64"
                },
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Log"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error on get all Logs!"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Publication": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "integer",
          "format": "int64"
        },
        "identifier": {
          "type": "string"
        },
        "datestamp": {
          "type": "string"
        },
        "metadata": {
          "type": "object",
          "properties": {
            "title": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "creator": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "subject": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "description": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "identifier": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "publisher": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "contributor": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "date": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "type": {
              "type": "string"
            },
            "lang": {
              "type": "string"
            },
            "rights": {
              "type": "string"
            }
          }
        },
        "data_provider": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "name": {
              "type": "string"
            },
            "url": {
              "type": "string"
            }
          }
        }
      }
    },
    "Log": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "integer"
        },
        "start_date": {
          "type": "string"
        },
        "end_date": {
          "type": "string"
        },
        "total_records": {
          "type": "integer"
        },
        "data_provider": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string"
            },
            "name": {
              "type": "string"
            },
            "url": {
              "type": "string"
            }
          }
        }
      }
    },
...
  }
}